further modifications

+++ dotfiles/install.sh
@@ -50,8 +50,19 @@ link_files ()
 
 copy_files () 
 {
-    cp -r $1 $2
-    success "Copied $1 to $2"
+    info "Copying $1 to $2"
+    if [ -f $2 -o -d $2 ]
+    then
+        info "$2 exists. Replacing old version of $2 with new one"
+        rm -rf $2
+        cp -r $1 $2
+	success "Copied $1 to $2"
+    else
+        cp -r $1 $2
+	success "Copied $1 to $2"
+    fi
+
+    
 }
 
 symlink_dotfiles () 
@@ -71,25 +82,18 @@ symlink_dotfiles ()
     done
 }
 
-#copy_dotfiles () 
-#{
-#    info 'Copying dotfiles to the home directory'
-#    for source in `find $DIR -maxdepth 2 -name \*.link`
-#    do
-#	dest="$HOME/.`basename \"${source%.*}\"`"
-#	if [ -f $dest -o -d $dest ]
-#	then
-#	    info "$dest exists. Replacing old version of $dest with new one"
-#	    rm -rf $dest
-#	    copy_files $source $dest
-#	else
-#	    copy_files $source $dest
-#	fi
-#    done
-#}
+# Copy git-prompt.sh (for git version control information in prompt)
+source="$HOME/Software/system-config/dotfiles/git/git-prompt.sh"
+dest="$HOME/.git-prompt.sh"
+copy_files $source $dest
 
-# Run functions
-set -e
-#copy_dotfiles
+# Copy tex.vim (for vim-latex modifications)
+source="$HOME/Software/system-config/dotfiles/vim-latex/tex.vim"
+dest="$HOME/.vim/bundle/vim-latex-1.8.23/ftplugin/tex.vim"
+copy_files $source $dest
+
+# Link dotfiles functions
 symlink_dotfiles
+
+# Source .bash_profile
 source $HOME/.bash_profile
